+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.13 (not yet released)
 * Add 'Create relationship (with relationship type parameter)' action
 * Add 'Subscribe to mailing list' action
 * Add 'Confirm mailing list subscription' action
+* Add 'Link contribution to participant' action
 * Changed Upload Custom File Field and Add Attachment action so it also accepts a URL for the attachment.
 
 Version 1.12